v0.6.8

+++ puphelpdesk/helpdesk/api/views/admin/Charter/charter.py
@@ -3,6 +3,9 @@ from rest_framework.decorators import api_view
 from api.serializers import CharterSerializer, CharterStepsSerializer
 from api.models import Charter, CharterSteps
 
+# For Searching Query
+from django.db.models import Q
+
 @api_view(['POST'])
 def adminAddCharter(request):
     if request.user.is_anonymous or not request.user.is_admin:
@@ -122,8 +125,12 @@ def adminDeleteCharter(request, charter_Id):
     else:
         if request.method == "DELETE":
             charter = Charter.objects.get(pk=charter_Id)
-            charter.delete()
-            return Response({"message": "Delete Charter Success"})
+            checkstep = CharterSteps.objects.all().filter(charter_Id=charter_Id).exists()
+            if checkstep:
+                return Response({"message": "Step Exist"})
+            else:
+                charter.delete()
+                return Response({"message": "Delete Charter Success"})
         return Response({"message": "Delete Charter Error"})
     
 ## Charter Step Section
@@ -212,3 +219,24 @@ def adminDeleteCharterStep(request, step_Id):
             charter.delete()
             return Response({"message": "Delete Charter Step Success"})
         return Response({"message": "Delete Charter Step Error"})
+    
+@api_view(['POST'])
+def adminSearchCharter(request, charter_Keyword):
+    if request.user.is_anonymous or not request.user.is_admin:
+        return Response({"message": "Not Authenticated"})
+    else:
+        if request.method == "POST":
+            charter_Keyword = request.POST.get('charter_Keyword')
+            data = Charter.objects.filter(
+                Q(charter_Avail__icontains=charter_Keyword) |
+                Q(charter_Classification__icontains=charter_Keyword) |
+                Q(charter_Description__icontains=charter_Keyword) |
+                Q(charter_Office__icontains=charter_Keyword) |
+                Q(charter_Requirements__icontains=charter_Keyword) |
+                Q(charter_Secure__icontains=charter_Keyword) |
+                Q(charter_Title__icontains=charter_Keyword) |
+                Q(charter_Transaction__icontains=charter_Keyword)
+            ).order_by('date_Created')
+            serializer = CharterSerializer(data, many=True)
+            return Response(serializer.data)
+        return Response({"message": "Get Charter Error"})
